Create LICENSE
As we scale up, we need to consider the possibility of people using our Timeout roles for their own purpose. This is fine, but we need to set up licensing. As our original source has granted their work an MIT license, we are free to choose whatever license we want.

Why have I chosen this license?

I strongly believe in the core tenets of free and open source software. This license enforces this in a more restrictive way compared to the MIT license by requiring that subsequent iterations have this very license. This preserves the open source aspect as more people explore this repository.